+++ hpxml-measures/HPXMLtoOpenStudio/resources/defaults.rb
@@ -654,16 +654,30 @@ module Defaults
     # Conductivity/diffusivity values come from https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4813881 (with the exception of "unknown")
     if hpxml_bldg.site.ground_conductivity.nil? && hpxml_bldg.site.ground_diffusivity.nil?
       case hpxml_bldg.site.soil_type
+      when HPXML::SiteSoilTypeUnknown, HPXML::SiteSoilTypeOther
+        case hpxml_bldg.site.moisture_type
+        when HPXML::SiteSoilMoistureTypeDry
+          hpxml_bldg.site.ground_conductivity = 0.3680 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0120 # ft^2/hr
+        when HPXML::SiteSoilMoistureTypeMixed
+          hpxml_bldg.site.ground_conductivity = 1.0000 # Btu/hr-ft-F, ANSI/RESNET/ICC 301-2022 Addendum C
+          hpxml_bldg.site.ground_diffusivity = 0.0208 # ft^2/hr
+        when HPXML::SiteSoilMoistureTypeWet
+          hpxml_bldg.site.ground_conductivity = 1.6320 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0296 # ft^2/hr
+        end
+        hpxml_bldg.site.ground_conductivity_isdefaulted = true
+        hpxml_bldg.site.ground_diffusivity_isdefaulted = true
       when HPXML::SiteSoilTypeSand
         if hpxml_bldg.site.moisture_type == HPXML::SiteSoilMoistureTypeDry
           hpxml_bldg.site.ground_conductivity = 0.2311 # Btu/hr-ft-F
           hpxml_bldg.site.ground_diffusivity = 0.0097 # ft^2/hr
+        elsif hpxml_bldg.site.moisture_type == HPXML::SiteSoilMoistureTypeMixed
+          hpxml_bldg.site.ground_conductivity = 0.8088 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0210 # ft^2/hr
         elsif hpxml_bldg.site.moisture_type == HPXML::SiteSoilMoistureTypeWet
           hpxml_bldg.site.ground_conductivity = 1.3865 # Btu/hr-ft-F
           hpxml_bldg.site.ground_diffusivity = 0.0322 # ft^2/hr
-        elsif hpxml_bldg.site.moisture_type == HPXML::SiteSoilMoistureTypeMixed
-          hpxml_bldg.site.ground_conductivity = ((0.2311 + 1.3865) / 2.0).round(4) # Btu/hr-ft-F
-          hpxml_bldg.site.ground_diffusivity = ((0.0097 + 0.0322) / 2.0).round(4) # ft^2/hr
         end
         hpxml_bldg.site.ground_conductivity_isdefaulted = true
         hpxml_bldg.site.ground_diffusivity_isdefaulted = true
@@ -672,18 +686,27 @@ module Defaults
         when HPXML::SiteSoilMoistureTypeDry
           hpxml_bldg.site.ground_conductivity = 0.2889 # Btu/hr-ft-F
           hpxml_bldg.site.ground_diffusivity = 0.0120 # ft^2/hr
+        when HPXML::SiteSoilMoistureTypeMixed
+          hpxml_bldg.site.ground_conductivity = 0.6355 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0157 # ft^2/hr
         when HPXML::SiteSoilMoistureTypeWet
           hpxml_bldg.site.ground_conductivity = 0.9821 # Btu/hr-ft-F
           hpxml_bldg.site.ground_diffusivity = 0.0194 # ft^2/hr
-        when HPXML::SiteSoilMoistureTypeMixed
-          hpxml_bldg.site.ground_conductivity = ((0.2889 + 0.9821) / 2.0).round(4) # Btu/hr-ft-F
-          hpxml_bldg.site.ground_diffusivity = ((0.0120 + 0.0194) / 2.0).round(4) # ft^2/hr
         end
         hpxml_bldg.site.ground_conductivity_isdefaulted = true
         hpxml_bldg.site.ground_diffusivity_isdefaulted = true
       when HPXML::SiteSoilTypeLoam
-        hpxml_bldg.site.ground_conductivity = 1.2132 # Btu/hr-ft-F
-        hpxml_bldg.site.ground_diffusivity = 0.0353 # ft^2/hr
+        case hpxml_bldg.site.moisture_type
+        when HPXML::SiteSoilMoistureTypeDry
+          hpxml_bldg.site.ground_conductivity = 0.4465 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0203 # ft^2/hr
+        when HPXML::SiteSoilMoistureTypeMixed
+          hpxml_bldg.site.ground_conductivity = 1.2132 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0353 # ft^2/hr
+        when HPXML::SiteSoilMoistureTypeWet
+          hpxml_bldg.site.ground_conductivity = 1.9799 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0502 # ft^2/hr
+        end
         hpxml_bldg.site.ground_conductivity_isdefaulted = true
         hpxml_bldg.site.ground_diffusivity_isdefaulted = true
       when HPXML::SiteSoilTypeGravel
@@ -691,18 +714,13 @@ module Defaults
         when HPXML::SiteSoilMoistureTypeDry
           hpxml_bldg.site.ground_conductivity = 0.2311 # Btu/hr-ft-F
           hpxml_bldg.site.ground_diffusivity = 0.0097 # ft^2/hr
+        when HPXML::SiteSoilMoistureTypeMixed
+          hpxml_bldg.site.ground_conductivity = 0.6355 # Btu/hr-ft-F
+          hpxml_bldg.site.ground_diffusivity = 0.0194 # ft^2/hr
         when HPXML::SiteSoilMoistureTypeWet
           hpxml_bldg.site.ground_conductivity = 1.0399 # Btu/hr-ft-F
           hpxml_bldg.site.ground_diffusivity = 0.0291 # ft^2/hr
-        when HPXML::SiteSoilMoistureTypeMixed
-          hpxml_bldg.site.ground_conductivity = ((0.2311 + 1.0399) / 2.0).round(4) # Btu/hr-ft-F
-          hpxml_bldg.site.ground_diffusivity = ((0.0097 + 0.0291) / 2.0).round(4) # ft^2/hr
         end
-        hpxml_bldg.site.ground_conductivity_isdefaulted = true
-        hpxml_bldg.site.ground_diffusivity_isdefaulted = true
-      when HPXML::SiteSoilTypeUnknown
-        hpxml_bldg.site.ground_conductivity = 1.0 # ANSI/RESNET/ICC 301-2022 Addendum C
-        hpxml_bldg.site.ground_diffusivity = 0.0208
         hpxml_bldg.site.ground_conductivity_isdefaulted = true
         hpxml_bldg.site.ground_diffusivity_isdefaulted = true
       end
@@ -4760,17 +4778,16 @@ module Defaults
           plug_load.monthly_multipliers_isdefaulted = true
         end
       when HPXML::PlugLoadTypeTelevision
-        default_annual_kwh, default_sens_frac, default_lat_frac = get_televisions_values(cfa, nbeds, n_occ, unit_type)
         if plug_load.kwh_per_year.nil?
-          plug_load.kwh_per_year = default_annual_kwh
+          plug_load.kwh_per_year = get_televisions_values(cfa, nbeds, n_occ, unit_type)
           plug_load.kwh_per_year_isdefaulted = true
         end
         if plug_load.frac_sensible.nil?
-          plug_load.frac_sensible = default_sens_frac
+          plug_load.frac_sensible = 1.0
           plug_load.frac_sensible_isdefaulted = true
         end
         if plug_load.frac_latent.nil?
-          plug_load.frac_latent = default_lat_frac
+          plug_load.frac_latent = 0.0
           plug_load.frac_latent_isdefaulted = true
         end
         schedules_file_includes_plug_loads_tv = (schedules_file.nil? ? false : schedules_file.includes_col_name(SchedulesFile::Columns[:PlugLoadsTV].name))
@@ -4787,7 +4804,7 @@ module Defaults
           plug_load.monthly_multipliers_isdefaulted = true
         end
       when HPXML::PlugLoadTypeElectricVehicleCharging
-        default_annual_kwh = get_electric_vehicle_charging_annual_energy
+        default_annual_kwh = get_electric_vehicle_charging_annual_energy(n_occ)
         if plug_load.kwh_per_year.nil?
           plug_load.kwh_per_year = default_annual_kwh
           plug_load.kwh_per_year_isdefaulted = true
@@ -7474,7 +7491,10 @@ module Defaults
   # @param unit_type [String] Type of dwelling unit (HXPML::ResidentialTypeXXX)
   # @return [Array<Double, Double, Double>] Plug loads annual use (kWh), sensible/latent fractions
   def self.get_residual_mels_values(cfa, n_occ = nil, unit_type = nil)
-    if n_occ.nil? # Asset calculation
+    if n_occ == 0
+      # Operational calculation w/ zero occupants, zero out energy use
+      annual_kwh = 0.0
+    elsif n_occ.nil? # Asset calculation
       # ANSI/RESNET/ICC 301
       annual_kwh = 0.91 * cfa
     else # Operational calculation
@@ -7495,15 +7515,18 @@ module Defaults
     return annual_kwh, frac_sens, frac_lat
   end
 
-  # Gets the default television energy use and sensible/latent fractions.
+  # Gets the default television energy use.
   #
   # @param cfa [Double] Conditioned floor area in the dwelling unit (ft2)
   # @param nbeds [Integer] Number of bedrooms in the dwelling unit
   # @param n_occ [Double] Number of occupants in the dwelling unit
   # @param unit_type [String] Type of dwelling unit (HXPML::ResidentialTypeXXX)
-  # @return [Array<Double, Double, Double>] Television annual use (kWh), sensible/latent fractions
+  # @return [Double] Television annual use (kWh)
   def self.get_televisions_values(cfa, nbeds, n_occ = nil, unit_type = nil)
-    if n_occ.nil? # Asset calculation
+    if n_occ == 0
+      # Operational calculation w/ zero occupants, zero out energy use
+      annual_kwh = 0.0
+    elsif n_occ.nil? # Asset calculation
       # ANSI/RESNET/ICC 301
       annual_kwh = 413.0 + 69.0 * nbeds
     else # Operational calculation
@@ -7524,10 +7547,7 @@ module Defaults
         annual_kwh = 99.9 + 129.6 * n_occ + 0.21 * cfa
       end
     end
-    frac_lost = 0.0
-    frac_sens = (1.0 - frac_lost) * 1.0
-    frac_lat = 1.0 - frac_sens - frac_lost
-    return annual_kwh, frac_sens, frac_lat
+    return annual_kwh
   end
 
   # Gets the default pool pump annual energy use.
@@ -7538,6 +7558,11 @@ module Defaults
   # @param unit_type [String] Type of dwelling unit (HXPML::ResidentialTypeXXX)
   # @return [Double] Annual energy use (kWh/yr)
   def self.get_pool_pump_annual_energy(cfa, nbeds, n_occ, unit_type)
+    if n_occ == 0
+      # Operational calculation w/ zero occupants, zero out energy use
+      return 0.0
+    end
+
     nbeds_eq = Defaults.get_equivalent_nbeds(nbeds, n_occ, unit_type)
 
     return 158.6 / 0.070 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0)
@@ -7558,13 +7583,23 @@ module Defaults
     load_value = nil
     if [HPXML::HeaterTypeElectricResistance, HPXML::HeaterTypeHeatPump].include? type
       load_units = HPXML::UnitsKwhPerYear
-      load_value = 8.3 / 0.004 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # kWh/yr
-      if type == HPXML::HeaterTypeHeatPump
-        load_value /= 5.0 # Assume seasonal COP of 5.0 per https://www.energy.gov/energysaver/heat-pump-swimming-pool-heaters
+      if n_occ == 0
+        # Operational calculation w/ zero occupants, zero out energy use
+        load_value = 0.0
+      else
+        load_value = 8.3 / 0.004 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # kWh/yr
+        if type == HPXML::HeaterTypeHeatPump
+          load_value /= 5.0 # Assume seasonal COP of 5.0 per https://www.energy.gov/energysaver/heat-pump-swimming-pool-heaters
+        end
       end
     elsif type == HPXML::HeaterTypeGas
       load_units = HPXML::UnitsThermPerYear
-      load_value = 3.0 / 0.014 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # therm/yr
+      if n_occ == 0
+        # Operational calculation w/ zero occupants, zero out energy use
+        load_value = 0.0
+      else
+        load_value = 3.0 / 0.014 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # therm/yr
+      end
     end
     return load_units, load_value
   end
@@ -7577,6 +7612,11 @@ module Defaults
   # @param unit_type [String] Type of dwelling unit (HXPML::ResidentialTypeXXX)
   # @return [Double] Annual energy use (kWh/yr)
   def self.get_permanent_spa_pump_annual_energy(cfa, nbeds, n_occ, unit_type)
+    if n_occ == 0
+      # Operational calculation w/ zero occupants, zero out energy use
+      return 0.0
+    end
+
     nbeds_eq = Defaults.get_equivalent_nbeds(nbeds, n_occ, unit_type)
 
     return 59.5 / 0.059 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # kWh/yr
@@ -7597,21 +7637,37 @@ module Defaults
     load_value = nil
     if [HPXML::HeaterTypeElectricResistance, HPXML::HeaterTypeHeatPump].include? type
       load_units = HPXML::UnitsKwhPerYear
-      load_value = 49.0 / 0.048 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # kWh/yr
-      if type == HPXML::HeaterTypeHeatPump
-        load_value /= 5.0 # Assume seasonal COP of 5.0 per https://www.energy.gov/energysaver/heat-pump-swimming-pool-heaters
+      if n_occ == 0
+        # Operational calculation w/ zero occupants, zero out energy use
+        load_value = 0.0
+      else
+        load_value = 49.0 / 0.048 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # kWh/yr
+        if type == HPXML::HeaterTypeHeatPump
+          load_value /= 5.0 # Assume seasonal COP of 5.0 per https://www.energy.gov/energysaver/heat-pump-swimming-pool-heaters
+        end
       end
     elsif type == HPXML::HeaterTypeGas
       load_units = HPXML::UnitsThermPerYear
-      load_value = 0.87 / 0.011 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # therm/yr
+      if n_occ == 0
+        # Operational calculation w/ zero occupants, zero out energy use
+        load_value = 0.0
+      else
+        load_value = 0.87 / 0.011 * (0.5 + 0.25 * nbeds_eq / 3.0 + 0.25 * cfa / 1920.0) # therm/yr
+      end
     end
     return load_units, load_value
   end
 
   # Gets the default electric vehicle charging annual energy use.
   #
+  # @param n_occ [Double] Number of occupants in the dwelling unit
   # @return [Double] Annual energy use (kWh/yr)
-  def self.get_electric_vehicle_charging_annual_energy()
+  def self.get_electric_vehicle_charging_annual_energy(n_occ)
+    if n_occ == 0
+      # Operational calculation w/ zero occupants, zero out energy use
+      return 0.0
+    end
+
     ev_charger_efficiency = 0.9
     ev_battery_efficiency = 0.9